Optimization and Class Structure Overhaul Made demonstrations that use demographics their own subclass Optimized including and excluding demo by precomputing them
